pyproject.toml
[project]
name = "py-ai-translator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic==1.15.1",
    "asyncpg==0.30.0",
    "crawl4ai==0.5.0.post4",
    "fastapi[all]==0.115.11",
    "loguru==0.7.3",
    "psycopg==3.2.6",
    "pydantic-ai-slim[logfire,openai]==0.0.36",
    "pydantic-settings==2.8.1",
    "sqlalchemy==2.0.39",
]
